Move summary display functions to cmd, replace returning verr to using atomicerr, change type of values to map
